requirements.txt
requests>=2.21.0
beautifulsoup4>=4.7.1
pyyaml>=3.13
